modules: mod_alloc: mod_alloc size and alignment parameter to size_t
Change mod_alloc_align and friends size and alignment parameters type
to size_t.

+++ src/include/sof/audio/module_adapter/module/generic.h
@@ -188,9 +188,9 @@ struct module_processing_data {
 /*****************************************************************************/
 int module_load_config(struct comp_dev *dev, const void *cfg, size_t size);
 int module_init(struct processing_module *mod);
-void *mod_alloc_align(struct processing_module *mod, uint32_t size, uint32_t alignment);
-void *mod_alloc(struct processing_module *mod, uint32_t size);
-void *mod_zalloc(struct processing_module *mod, uint32_t size);
+void *mod_alloc_align(struct processing_module *mod, size_t size, size_t alignment);
+void *mod_alloc(struct processing_module *mod, size_t size);
+void *mod_zalloc(struct processing_module *mod, size_t size);
 int mod_free(struct processing_module *mod, const void *ptr);
 #if CONFIG_COMP_BLOB
 struct comp_data_blob_handler *mod_data_blob_handler_new(struct processing_module *mod);
